Separate network portion into a new file
